Renamed Widget to MainWindow
